+++ cursos/src/main/java/com/example/cursos/repositories/CourseRepository.java
@@ -2,6 +2,6 @@ package com.example.cursos.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.cursos.models.Course;
-public interface CourseRepository extends JpaRepository<Course, Long> {
+public interface CourseRepository extends JpaRepository<Course, Integer> {
 
 }
